back: [fix] resolve lua issue
Bug: This issue is a memory issue caused on Wiktionaries side. It sometimes happens on pages with a lot of entries.

Fix: Not the best fix, ideally we find a way resolve it using Wiktionary however to save time, we'll use this archaic method. So we assume that BECAUSE there's a lot of entries for the word we're searching, the word exists in many languages and therefore there isn't a translation for it.