[feat] improved homonym support
To make things more consistent storeDuolingoData() and storeData() have been modified to use the same data structure as local storage and data from the backend. This simplifies working with the data and forgoes the need to alter data in order to use storeData().

We now also store a words Duolingo ID to make comparisons between locally stored words and new words easier to identify.

There is an issue with some words being the exact same but duolingo has different ids (e.g. "por favor"). There are two issues with this, 1. is that for whatever reason, even though it should be, the code isn't storing both ids (this sounds like a small bug I can fix). The other issue is how I'm going to handle having two of the exact same words