back: [fix] change processWord() to handle arrays
Because of homonyms, things have had to be changed. The structure of data now has an array of objects associated to a word. To handle these array of objects things have changed AND some code has been added so that a single word is searched only once and the right words from the client are matched with the words from wiktionary