back: [fix] many client infinitives searched
Bug: Another issue that arose with the changes that came with homonym support was client infinitives would be searched multiple times. The infinitive would have n objects in it's array, where n is the amount of words associated with the infinitive.

Fix: This is fixed using an object instead of array to store the promises. If the key exists in the object then it won't add a new promises, even if that condition wasn't there, because we use objects, the key would just be overridden and still leave a single promise.