Game: [feat] basic logic
The word in the learning language shows up, you input the translation, if it's right the word you don't have to type it again, if it's wrong, you'll be tested on it again. There's also a restart button, to reset the game.

Forget the styling, but there are things that still need to be added.